Reduce the number of bytes in a chunk header from 48 to 40 by packing chunk_type and num_records in a single word.
This changes the file format in an incompatible way.

PiperOrigin-RevId: 195097709

+++ riegeli/chunk_encoding/types.h
@@ -20,7 +20,7 @@
 namespace riegeli {
 
 // These values are frozen in the file format.
-enum class ChunkType : uint64_t {
+enum class ChunkType : uint8_t {
   kFileSignature = 's',
   kPadding = 'p',
   kSimple = 'r',
